+++ packages/browser/src/login/oidc/OidcLoginHandler.ts
@@ -39,6 +39,7 @@ import {
   IOidcOptions,
   IStorageUtility,
   ConfigurationError,
+  LoginResult,
 } from "@inrupt/solid-client-authn-core";
 
 import { IClient } from "@inrupt/oidc-client-ext";
@@ -72,7 +73,7 @@ export default class OidcLoginHandler implements ILoginHandler {
     return hasIssuer(options) && hasRedirectUrl(options);
   }
 
-  async handle(options: ILoginOptions): Promise<void> {
+  async handle(options: ILoginOptions): Promise<LoginResult> {
     if (!hasIssuer(options)) {
       throw new ConfigurationError(
         `OidcLoginHandler requires an OIDC issuer: missing property 'oidcIssuer' in ${JSON.stringify(
@@ -148,6 +149,6 @@ export default class OidcLoginHandler implements ILoginHandler {
     };
 
     // Call proper OIDC Handler
-    await this.oidcHandler.handle(OidcOptions);
+    return this.oidcHandler.handle(OidcOptions);
   }
 }