+++ packages/node/src/login/oidc/OidcLoginHandler.spec.ts
@@ -60,16 +60,17 @@ describe("OidcLoginHandler", () => {
       ).resolves.toEqual(false);
     });
 
-    it("cannot handle options without an redirect url", async () => {
-      const handler = getInitialisedHandler();
-      await expect(
-        handler.canHandle({
-          sessionId: "mySession",
-          tokenType: "DPoP",
-          oidcIssuer: "https://my.idp/",
-        })
-      ).resolves.toEqual(false);
-    });
+    // TODO: Move this to appropriate handlers (auth code, implicit)
+    // it("cannot handle options without an redirect url", async () => {
+    //   const handler = getInitialisedHandler();
+    //   await expect(
+    //     handler.canHandle({
+    //       sessionId: "mySession",
+    //       tokenType: "DPoP",
+    //       oidcIssuer: "https://my.idp/",
+    //     })
+    //   ).resolves.toEqual(false);
+    // });
 
     it("can handle options with both a redirect url and an issuer", async () => {
       const handler = getInitialisedHandler();
@@ -96,16 +97,17 @@ describe("OidcLoginHandler", () => {
       ).rejects.toThrow("OidcLoginHandler requires an OIDC issuer");
     });
 
-    it("throws if config misses a redirect URL", async () => {
-      const handler = getInitialisedHandler();
-      await expect(
-        handler.handle({
-          sessionId: "mySession",
-          tokenType: "DPoP",
-          oidcIssuer: "https://my.idp/",
-        })
-      ).rejects.toThrow("OidcLoginHandler requires a redirect URL");
-    });
+    // TODO: Move this to appropriate handlers (auth code, implicit)
+    // it("throws if config misses a redirect URL", async () => {
+    //   const handler = getInitialisedHandler();
+    //   await expect(
+    //     handler.handle({
+    //       sessionId: "mySession",
+    //       tokenType: "DPoP",
+    //       oidcIssuer: "https://my.idp/",
+    //     })
+    //   ).rejects.toThrow("OidcLoginHandler requires a redirect URL");
+    // });
 
     it("performs DCR if client ID and secret aren't specified", async () => {
       const { oidcHandler } = defaultMocks;